pyproject.toml
[project]
name = "finguard-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.43.0",
    "langchain>=1.2.16",
    "langchain-aws>=1.4.5",
    "langchain-community>=0.4.1",
    "langgraph>=1.1.10",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.57.0",
]
